Fix more compiler warnings

+++ Engine/Rendering/TMView.mm
@@ -109,7 +109,7 @@
     if (m_pControls->empty())
         return nil;
 
-    int curSize = m_pControls->size();
+    unsigned long curSize = m_pControls->size();
 
     for (int i = 0; i < curSize; ++i)
     {
